+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Unreleased
 ----------
+- Added additional generic `Message` type to `Ui`, `Widget`, and
+  more structs & traits
 - Converted all event handlers to async functions
   - Added `async-trait` dependency in version `0.1`